creacion de los Accessor y Mutator para el modelo de productos en el precio de venta

+++ database/migrations/2025_03_11_225700_create_productos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('proveedor_id')->constrained('proveedors')->onDelete('cascade');
             $table->string('nombre_producto');
-            $table->decimal('precio_venta',12, 3);
+            $table->bigInteger('precio_venta');
             $table->timestamps();
         });
     }
